ffnerd v0.1.0 FFNerd.DepthChart

Provides functions to work with Fantasy Football Nerd’s Depth Chart resources.

More info at: http://www.fantasyfootballnerd.com/fantasy-football-api#depth-charts

Functions

find(team, client)

Return a team’s depth chart record by team code.

Examples

FFNerd.DepthChart.find “SEA”, client

find(team, position, client)

Return a team’s position depth chart record by team code and position.

Examples

FFNerd.DepthChart.find “SEA”, “QB”, client

find(team, position, depth, client)

Return a depth chart player record by team code, position, and depth.

Examples

FFNerd.DepthChart.find “SEA”, “QB”, 1, client

list(client)

Return a list of depth chart records.

Examples

FFNerd.DepthChart.list client

new(map)

Convert raw api data to DepthChart struct.