ffnerd v0.1.0 FFNerd.Injury

Provides functions to work with Fantasy Football Nerd’s Injury resources.

More info at: http://www.fantasyfootballnerd.com/fantasy-football-api#injuries

Functions

find(id, team, client)

Return a single injury record by player id and team code.

Examples

FFNerd.Injury.find 1465, “SEA”, client

find(id, team, week, client)

Return a single injury record by player id, team code, and week.

Examples

FFNerd.Injury.find 1465, “SEA”, 17, client

list(client)

Return a list of injury records.

Examples

FFNerd.Injury.list client

list(week, client)

Return a list of current week injury records by team.

Examples

FFNerd.Injury.list “SEA”, client

list(team, week, client)

Return a list of injury records by team and week.

Examples

FFNerd.Injury.list “SEA”, 17, client

new(map)

Convert raw api data to Player.Injury struct.