ffnerd v0.0.1 FFNerd.Player.AuctionValues

Provides functions to work with Fantasy Football Nerd’s Auction Values resources.

More info at: http://www.fantasyfootballnerd.com/fantasy-football-api#auction

Functions

find(id, client)

Return a single auction value record by person id.

Examples

FFNerd.Player.AuctionValues.find 259, client

FFNerd.Player.AuctionValues.find :ppr, 259, client

find(atom, id, client)

list(client)

Return a list of auction value records.

Examples

FFNerd.Player.AuctionValues.list client

FFNerd.Player.AuctionValues.list :ppr, client

list(position, client)

Return a list of auction value records by position.

Examples

FFNerd.Player.AuctionValues.list “QB”, client

FFNerd.Player.AuctionValues.list :ppr, “QB”, client

list(atom, position, client)

new(map_or_kwlist, opts \\ [])

Specs

new(ExConstructor.map_or_kwlist, Keyword.t) :: %FFNerd.Player.AuctionValues{avg_price: term, display_name: term, max_price: term, min_price: term, player_id: term, position: term, ppr: term, team: term}